requirements.txt

# Slack API
slack-bolt>=1.15.0
slack-sdk>=3.19.0

# Todoist API
todoist-api-python>=2.0.0

# Google Calendar API
google-api-python-client>=2.70.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0

# OpenAI API
openai>=1.0.0

# Anthropic Claude API
anthropic>=0.24.0

# Utilities
python-dotenv>=0.21.0
pytz>=2022.7
requests>=2.28.0
